whisper-server/requirements.txt
flask>=3.0.0,<4.0.0
flask-cors>=4.0.0,<5.0.0
openai-whisper>=20240930
gunicorn>=22.0.0,<23.0.0
